pyproject.toml
[tool.poetry]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = ["RenatoAkira <renatoakira@hotmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "*"
psycopg2-binary = "*"
gunicorn = "*"
flask-bcrypt = "*"
flask-marshmallow = "*"
marshmallow-sqlalchemy = "*"
apispec-webframeworks = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
